Attempt to work around cases where numpy.linalg.lstsq fails by computing Moore-Penrose pseudoinverse explicitly
- numpy.linalg.lstsq is the preferred way to solev equations of the form A.x = b, where A may not have an inverse in the strict sense, because it is supposedly equivalent to, but numerically more stable than, computing the pseudo-inverse and left-multiplying. Generally we try to follow this policy. However, in some cases, numpy appears not to converge when solving the least-squares equation.

(it isn't actually clear (to me) what numpy does internally here; some discussions of the scipy equivalent function seem to imply it first computes an SVD decomposition. But it would be odd if this were the case, since then it is hard to see why numpy.linalg.lstsq would actually have different stability properties to numpy.linalg.pinv. So presumably this is *not* what happens - at some point, perhaps check the numpy sources?)

- when numpy.linalg.lstsq fails, we now fall back to an explicit computation of the Moore-Penrose pseudoinverse. In numerical experiments this seems to work OK, at least so far